Try fixing lock tests on Windows
