+++ lib/scala/runtime-version-manager-test/src/test/scala/org/enso/distribution/locking/ThreadSafeFileLockManagerTest.scala
@@ -8,7 +8,10 @@ import org.enso.distribution.locking.{
 }
 
 import java.nio.file.Path
-import org.enso.runtimeversionmanager.test.NativeTestHelper
+import org.enso.runtimeversionmanager.test.{
+  NativeTestHelper,
+  TestableThreadSafeFileLockManager
+}
 import org.enso.testkit.{TestSynchronizer, WithTemporaryDirectory}
 import org.scalatest.OptionValues
 import org.scalatest.concurrent.TimeLimitedTests
@@ -26,28 +29,6 @@ class ThreadSafeFileLockManagerTest
     with NativeTestHelper {
 
   override val timeLimit: Span = 30.seconds
-
-  class TestableThreadSafeFileLockManager(locksRoot: Path)
-      extends ThreadSafeFileLockManager(locksRoot) {
-
-    /** A helper function that can be called by the test suite to release all file locks.
-      *
-      * It is only safe to call this function if it can be guaranteed that no
-      * locks created with this manager are still in scope.
-      *
-      * Normally all file locks are released automatically when the JVM exits,
-      * but as tests run within a single JVM, this is not the case and any
-      * dangling locks will cause problems when cleaning the temporary
-      * directory.
-      */
-    def releaseAllLocks(): Unit = {
-      localLocks.foreach { case (_, lock) =>
-        lock.fileLock.foreach(_.release())
-        lock.fileLock = None
-      }
-      localLocks.clear()
-    }
-  }
 
   private var testLocalLockManager: Option[TestableThreadSafeFileLockManager] =
     None
